pyproject.toml
[tool.poetry]
name = "gym-bud-app"
version = "0.1.0"
description = ""
authors = ["Rafael Herrero <rafaelgherrero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
requests = "^2.28.0"
SQLAlchemy = "^1.4.37"
uvicorn = "^0.17.6"
psycopg2-binary = "^2.9.1"
structlog = "^21.5.0"
python-jose = {version = "*", extras = ["cryptography"]}
passlib = {version = "*", extras = ["bcrypt"]}
python-multipart = "^0.0.5"
django = "^4.1.7"
djangorestframework = "^3.14.0"
coreapi = "^2.3.3"
drf-yasg = "^1.21.5"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
